serveur/requirements.txt
aniso8601==10.0.1
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
Flask==3.1.1
flask-cors==6.0.0
Flask-HTTPAuth==4.8.0
Flask-RESTful==0.3.10
Flask-WTF==1.2.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.3.0
paho-mqtt==2.1.0
pytz==2025.2
requests==2.32.3
six==1.17.0
urllib3==2.4.0
Werkzeug==3.1.3
WTForms==3.2.1
websockets==15.0.1
